serverless/code/requirements.txt
google-cloud-storage
functions-framework==3.*
Flask~=2.2.2